+++ PROGulky/Modules/ExcursionsList/View/ExcursionViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  Excursion.swift
+//  PROGulky
+//
+//  Created by Semyon Pyatkov on 06.11.2022.
+//
+
+import Foundation
+
+// MARK: - ExcursionViewModel
+
+struct ExcursionViewModel {
+    var image: String?
+    var title: String
+    var rating: Double
+    var parameters: String
+}
